v1.4.6.23
## 1.4.6.23 ##
=========
-- Moved "Or" items to Objective, using the symbols of "+ " but of the objective color.
-- Changed labels on Quest Steps.
        *  = Objective
        +  = Choice Objective (Or Objectives, or choice of location)
        ** = Optional Objective/Comments

I apologize, the game actually has it "Or" items labeled as objectives, I put them under a different category so they would stand out more in the tracker.  I did not take into account that some people could be hiding hidden, hints, or optional items.  I also discovered because of feedback from users, that the "Or" items are not only used when you have a choice of which NPC to talk to for example, it can also be used when you have a choice of where to go to complete this, or there are multiple locations and yo may only have to go to some of them not all of them.

So I moved these items back to being a Objective, but instead of being proceeded by a "*?" it is proceeded by a "+".   For example in the Madness in Shadowfen quest at the end you are to speak with Alvur Baren, it is showing up as he is actually in multiple location, you can actually turn the quest in at three different locations, as he is in Mages Guild in Elden Root, Wayrest, and Mournhold.  This is not limited to just quest turn ins but also cases where you have multiple locations but only have to get one or a few but not all locations.

I still want to mark this quest as something different especially when you do have the original intent of a choice between two people.  So these quests steps will now be proceeded with a "+", indicating you do have choices.  If actual independent choices show up, you only will need to complete one.  If only one step shows as a "+" then you have a choice where to obtain that objective.